Further clang-tidy fixes in header files

+++ src/lib/utils/alignment_buffer.h
@@ -31,7 +31,7 @@ namespace Botan {
  * input data is available in the BufferSlicer<>. This might result in some
  * performance overhead when using the must_be_deferred strategy.
  */
-enum class AlignmentBufferFinalBlock : size_t {
+enum class AlignmentBufferFinalBlock : uint8_t {
    is_not_special = 0,
    must_be_deferred = 1,
 };
